+++ src/interfaces/cli/types.ts
@@ -1,0 +1,7 @@
+export interface CommandEntry {
+  command: string;
+  output: string;
+  isStreaming?: boolean;
+  agentName: string;
+}
+
